+++ packages/react-ui/src/index.d.ts
@@ -42,6 +42,9 @@ import SelectField, { SelectFieldProps } from './components/atoms/SelectField';
 import UploadField, {
   UploadFieldProps
 } from './components/molecules/UploadField/UploadField';
+import UploadFieldBase, {
+  UploadFieldBaseProps
+} from './components/molecules/UploadField/UploadFieldBase';
 import AppBar, { AppBarProps } from './components/organisms/AppBar/AppBar';
 import LabelWithIndicator, {
   LabelWithIndicatorProps
@@ -104,6 +107,8 @@ export {
   SelectFieldProps,
   UploadField,
   UploadFieldProps,
+  UploadFieldBase,
+  UploadFieldBaseProps,
   AppBar,
   AppBarProps,
   LabelWithIndicator,