+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Not released
 
+- UploadField component: UI fixes and extract pure UI components [#817](https://github.com/CartoDB/carto-react/pull/817)
+
 ## 2.3
 
 ### 2.3.6 (2024-01-04)